TASK: Move filters to new files
